refactor(resourcecodec): consolidate codec configs into single file
Why this change was needed:
Having 11 separate files each containing a single small function
(5-9 lines) was excessive file proliferation. This made navigation
harder and added unnecessary cognitive overhead.

What changed:
- Created configs.go with all 11 resource codec config functions
- Removed 11 separate files (route.go, service.go, upstream.go,
  consumer.go, consumer_group.go, plugin_config.go, global_rule.go,
  plugin_metadata.go, proto.go, ssl.go, stream_route.go)

Problem solved:
Reduced from 11 files to 1 file while maintaining identical
functionality. Easier to view all codec configs at once and
identify patterns across resource types.
