--- src/apiserver/pkg/resourcecodec/consumer_group.go
@@ -1,8 +1,0 @@
-package resourcecodec
-
-func consumerGroupCodecConfig() resourceCodecConfig {
-	return resourceCodecConfig{
-		nameKey:     "name",
-		stripFields: []string{"id", "name"},
-	}
-}
