--- src/apiserver/pkg/resourcecodec/global_rule.go
@@ -1,8 +1,0 @@
-package resourcecodec
-
-func globalRuleCodecConfig() resourceCodecConfig {
-	return resourceCodecConfig{
-		nameKey:     "name",
-		stripFields: []string{"id", "name"},
-	}
-}
