--- src/apiserver/pkg/resourcecodec/plugin_config.go
@@ -1,8 +1,0 @@
-package resourcecodec
-
-func pluginConfigCodecConfig() resourceCodecConfig {
-	return resourceCodecConfig{
-		nameKey:     "name",
-		stripFields: []string{"id", "name"},
-	}
-}
