--- src/apiserver/pkg/resourcecodec/plugin_metadata.go
@@ -1,8 +1,0 @@
-package resourcecodec
-
-func pluginMetadataCodecConfig() resourceCodecConfig {
-	return resourceCodecConfig{
-		nameKey:     "name",
-		stripFields: []string{"id", "name"},
-	}
-}
