--- src/apiserver/pkg/resourcecodec/proto.go
@@ -1,8 +1,0 @@
-package resourcecodec
-
-func protoCodecConfig() resourceCodecConfig {
-	return resourceCodecConfig{
-		nameKey:     "name",
-		stripFields: []string{"id", "name"},
-	}
-}
