--- src/apiserver/pkg/resourcecodec/service.go
@@ -1,9 +1,0 @@
-package resourcecodec
-
-func serviceCodecConfig() resourceCodecConfig {
-	return resourceCodecConfig{
-		nameKey:           "name",
-		associationFields: []string{"upstream_id"},
-		stripFields:       []string{"id", "name", "upstream_id"},
-	}
-}
