--- src/apiserver/pkg/resourcecodec/ssl.go
@@ -1,8 +1,0 @@
-package resourcecodec
-
-func sslCodecConfig() resourceCodecConfig {
-	return resourceCodecConfig{
-		nameKey:     "name",
-		stripFields: []string{"id", "name"},
-	}
-}
