--- src/apiserver/pkg/resourcecodec/stream_route.go
@@ -1,9 +1,0 @@
-package resourcecodec
-
-func streamRouteCodecConfig() resourceCodecConfig {
-	return resourceCodecConfig{
-		nameKey:           "name",
-		associationFields: []string{"service_id", "upstream_id"},
-		stripFields:       []string{"id", "name", "service_id", "upstream_id"},
-	}
-}
